moved ulility function for home page to lib/home and reduce debouce time for search to 100ms
